fuck

+++ include/utils.h
@@ -6,7 +6,7 @@
 /*   By: evallee- <evallee-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/30 06:31:33 by niceguy           #+#    #+#             */
-/*   Updated: 2023/08/25 17:54:24 by evallee-         ###   ########.fr       */
+/*   Updated: 2023/08/25 18:54:00 by evallee-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 # include <stdlib.h>
 # include <stdbool.h>
 
-int32_t		ft_atoi(const char *str);
+long		ft_atoi(const char *str);
 uint64_t	get_time(uint64_t start);
 bool		ft_check_args(int argc, char **argv);
